requirements.txt
fastapi
uvicorn
python-pptx
python-multipart
openpyxl